+++ MacAProject/Model/Model.swift
@@ -87,3 +87,4 @@ extension CoffeeList{
 //  ]
     
 }
+
